refactor: extract saved hand export and stats
