+++ lib/screens/session_hands_screen.dart
@@ -6,6 +6,8 @@ import '../widgets/session_label_overlay.dart';
 
 import '../models/saved_hand.dart';
 import '../services/saved_hand_manager_service.dart';
+import '../services/saved_hand_export_service.dart';
+import '../services/saved_hand_stats_service.dart';
 import '../services/session_note_service.dart';
 import '../services/session_manager.dart';
 import '../widgets/saved_hand_tile.dart';
@@ -175,10 +177,10 @@ class _SessionHandsScreenState extends State<SessionHandsScreen> {
   }
 
   Future<void> _exportMarkdown(BuildContext context) async {
-    final manager = context.read<SavedHandManagerService>();
+    final exporter = context.read<SavedHandExportService>();
     final note = context.read<SessionNoteService>().noteFor(widget.sessionId);
-    final path =
-        await manager.exportSessionHandsMarkdown(widget.sessionId, note: note);
+    final path = await exporter.exportSessionHandsMarkdown(widget.sessionId,
+        note: note);
     if (path == null) return;
     await Share.shareXFiles([XFile(path)], text: 'session_${widget.sessionId}.md');
     if (context.mounted) {
@@ -189,10 +191,10 @@ class _SessionHandsScreenState extends State<SessionHandsScreen> {
   }
 
   Future<void> _exportPdf(BuildContext context) async {
-    final manager = context.read<SavedHandManagerService>();
+    final exporter = context.read<SavedHandExportService>();
     final note = context.read<SessionNoteService>().noteFor(widget.sessionId);
     final path =
-        await manager.exportSessionHandsPdf(widget.sessionId, note: note);
+        await exporter.exportSessionHandsPdf(widget.sessionId, note: note);
     if (path == null) return;
     await Share.shareXFiles([XFile(path)], text: 'session_${widget.sessionId}.pdf');
     if (context.mounted) {
@@ -205,12 +207,13 @@ class _SessionHandsScreenState extends State<SessionHandsScreen> {
   @override
   Widget build(BuildContext context) {
     final manager = context.watch<SavedHandManagerService>();
+    final stats = context.watch<SavedHandStatsService>();
     final hands = manager.hands
         .where((h) => h.sessionId == widget.sessionId)
         .toList()
       ..sort((a, b) => b.savedAt.compareTo(a.savedAt));
 
-    final sessionIds = manager.handsBySession().keys.toList()..sort();
+    final sessionIds = stats.handsBySession().keys.toList()..sort();
     final currentIndex = sessionIds.indexOf(widget.sessionId);
     final previousId =
         currentIndex > 0 ? sessionIds[currentIndex - 1] : null;
