+++ lib/services/saved_hand_manager_service.dart
@@ -1,17 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'dart:io';
 import 'dart:async';
-import 'package:path_provider/path_provider.dart';
-import 'package:archive/archive.dart';
-import 'package:pdf/widgets.dart' as pw;
-import 'package:pdf/pdf.dart';
-import 'package:csv/csv.dart';
 import 'package:uuid/uuid.dart';
-import 'package:permission_handler/permission_handler.dart';
-import 'package:share_plus/share_plus.dart';
 import '../helpers/date_utils.dart';
-import '../helpers/export_utils.dart';
 
 import '../models/saved_hand.dart';
 import '../models/v2/training_pack_template.dart';
@@ -23,25 +14,6 @@ import 'saved_hand_storage_service.dart';
 import 'cloud_sync_service.dart';
 
 import 'training_stats_service.dart';
-
-class _SessionStats {
-  _SessionStats(this.start, this.end, this.count, this.correct, this.incorrect,
-      {this.evAvg, this.icmAvg});
-
-  final DateTime start;
-  final DateTime end;
-  final int count;
-  final int correct;
-  final int incorrect;
-  final double? evAvg;
-  final double? icmAvg;
-
-  Duration get duration => end.difference(start);
-  double? get winrate {
-    final total = correct + incorrect;
-    return total > 0 ? correct / total * 100 : null;
-  }
-}
 
 class SavedHandManagerService extends ChangeNotifier {
   static SavedHandManagerService? _instance;
@@ -160,458 +132,6 @@ class SavedHandManagerService extends ChangeNotifier {
   }
 
   SavedHand? get lastHand => hands.isNotEmpty ? hands.last : null;
-
-  /// Export all saved hands to a Markdown file located in the
-  /// application documents directory. Returns the file path or `null`
-  /// if there are no saved hands.
-  Future<String?> exportAllHandsMarkdown() async {
-    if (hands.isEmpty) return null;
-    final buffer = StringBuffer();
-    for (final hand in hands) {
-      buffer.write(ExportUtils.handMarkdown(hand));
-    }
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/all_saved_hands.md');
-    await file.writeAsString(buffer.toString());
-    return file.path;
-  }
-
-  /// Export all saved hands to a PDF file located in the
-  /// application documents directory. Returns the file path or `null`
-  /// if there are no saved hands.
-  Future<String?> exportAllHandsPdf() async {
-    if (hands.isEmpty) return null;
-
-    final regularFont = await pw.PdfGoogleFonts.robotoRegular();
-    final boldFont = await pw.PdfGoogleFonts.robotoBold();
-
-    final pdf = pw.Document();
-
-    pdf.addPage(
-      pw.MultiPage(
-        pageFormat: PdfPageFormat.a4,
-        build: (context) {
-          return [
-            for (final hand in hands) ...ExportUtils.handPdfWidgets(
-              hand,
-              regularFont,
-              boldFont,
-            ),
-          ];
-        },
-      ),
-    );
-
-    final bytes = await pdf.save();
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/all_saved_hands.pdf');
-    await file.writeAsBytes(bytes);
-    await _shareFile(file);
-    return file.path;
-  }
-
-  /// Export statistics for all training sessions to a Markdown file named
-  /// `training_summary.md` located in the application documents directory.
-  /// [notes] should contain optional session notes keyed by session id.
-  Future<String?> exportAllSessionsMarkdown(Map<int, String> notes) async {
-    if (hands.isEmpty) return null;
-
-    String durToStr(Duration d) => ExportUtils.durationString(d);
-
-    final grouped = handsBySession().entries.toList()
-      ..sort((a, b) => a.key.compareTo(b.key));
-    final buffer = StringBuffer();
-
-    for (final entry in grouped) {
-      final id = entry.key;
-      final sessionHands = List<SavedHand>.from(entry.value)
-        ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-      if (sessionHands.isEmpty) continue;
-
-      final stats = _sessionStats(sessionHands);
-      final note = notes[id];
-
-      buffer.writeln('## Сессия $id');
-      buffer.writeln('- Дата: ${formatDateTime(stats.end)}');
-      buffer.writeln('- Длительность: ${durToStr(stats.duration)}');
-      buffer.writeln('- Раздач: ${stats.count}');
-      buffer.writeln('- Верно: ${stats.correct}');
-      buffer.writeln('- Ошибки: ${stats.incorrect}');
-      final winrate = stats.winrate;
-      if (winrate != null) {
-        buffer.writeln('- Winrate: ${winrate.toStringAsFixed(1)}%');
-      }
-      if (note != null && note.trim().isNotEmpty) {
-        buffer.writeln('- Заметка: ${note.trim()}');
-      }
-      buffer.writeln();
-      for (final hand in sessionHands) {
-        buffer.write(ExportUtils.handMarkdown(hand, level: 3));
-      }
-    }
-
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/training_summary.md');
-    await file.writeAsString(buffer.toString());
-    return file.path;
-  }
-
-  /// Export statistics for all training sessions to a PDF file named
-  /// `training_summary.pdf` located in the application documents directory.
-  /// [notes] should contain optional session notes keyed by session id.
-  Future<String?> exportAllSessionsPdf(Map<int, String> notes) async {
-    if (hands.isEmpty) return null;
-
-    String durToStr(Duration d) => ExportUtils.durationString(d);
-
-    final regularFont = await pw.PdfGoogleFonts.robotoRegular();
-    final boldFont = await pw.PdfGoogleFonts.robotoBold();
-
-    final grouped = handsBySession().entries.toList()
-      ..sort((a, b) => a.key.compareTo(b.key));
-
-    final pdf = pw.Document();
-    pdf.addPage(
-      pw.MultiPage(
-        pageFormat: PdfPageFormat.a4,
-        build: (context) {
-          return [
-            for (final entry in grouped) ...[
-              () {
-                final id = entry.key;
-                final sessionHands = List<SavedHand>.from(entry.value)
-                  ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-                final stats = _sessionStats(sessionHands);
-                final note = notes[id];
-
-                return [
-                  pw.Text('Сессия $id',
-                      style: pw.TextStyle(font: boldFont, fontSize: 20)),
-                  pw.SizedBox(height: 4),
-                  pw.Text('Дата: ${formatDateTime(stats.end)}',
-                      style: pw.TextStyle(font: regularFont)),
-                  pw.Text('Длительность: ${durToStr(stats.duration)}',
-                      style: pw.TextStyle(font: regularFont)),
-                  pw.Text('Раздач: ${stats.count} • Верно: ${stats.correct} • Ошибки: ${stats.incorrect}',
-                      style: pw.TextStyle(font: regularFont)),
-                  if (stats.winrate != null)
-                    pw.Text('Winrate: ${stats.winrate!.toStringAsFixed(1)}%',
-                        style: pw.TextStyle(font: regularFont)),
-                  if (note != null && note.trim().isNotEmpty)
-                    pw.Text('Заметка: ${note.trim()}',
-                        style: pw.TextStyle(font: regularFont)),
-                  pw.SizedBox(height: 8),
-                  for (final hand in sessionHands)
-                    ...ExportUtils.handPdfWidgets(
-                      hand,
-                      regularFont,
-                      boldFont,
-                      titleSize: 16,
-                    ),
-                  pw.Divider(),
-                ];
-              }(),
-            ],
-          ];
-        },
-      ),
-    );
-
-    final bytes = await pdf.save();
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/training_summary.pdf');
-    await file.writeAsBytes(bytes);
-    await _shareFile(file);
-    return file.path;
-  }
-
-  /// Export statistics for the specified training [sessionIds] to a Markdown
-  /// file named `training_summary_filtered.md` located in the application
-  /// documents directory. [notes] should contain optional session notes keyed
-  /// by session id.
-  Future<String?> exportSessionsMarkdown(
-      List<int> sessionIds, Map<int, String> notes) async {
-    if (sessionIds.isEmpty) return null;
-
-    String durToStr(Duration d) => ExportUtils.durationString(d);
-
-    final grouped = handsBySession();
-    final ids = List<int>.from(sessionIds)..sort();
-    final buffer = StringBuffer();
-
-    for (final id in ids) {
-      final sessionHands = grouped[id];
-      if (sessionHands == null || sessionHands.isEmpty) continue;
-      final handsList = List<SavedHand>.from(sessionHands)
-        ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-      final stats = _sessionStats(handsList);
-      final note = notes[id];
-
-      buffer.writeln('## Сессия $id');
-      buffer.writeln('- Дата: ${formatDateTime(stats.end)}');
-      buffer.writeln('- Длительность: ${durToStr(stats.duration)}');
-      buffer.writeln('- Раздач: ${stats.count}');
-      buffer.writeln('- Верно: ${stats.correct}');
-      buffer.writeln('- Ошибки: ${stats.incorrect}');
-      final winrate = stats.winrate;
-      if (winrate != null) {
-        buffer.writeln('- Winrate: ${winrate.toStringAsFixed(1)}%');
-      }
-      if (note != null && note.trim().isNotEmpty) {
-        buffer.writeln('- Заметка: ${note.trim()}');
-      }
-      buffer.writeln();
-      for (final hand in handsList) {
-        buffer.write(ExportUtils.handMarkdown(hand, level: 3));
-      }
-    }
-
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/training_summary_filtered.md');
-    await file.writeAsString(buffer.toString());
-    return file.path;
-  }
-
-  /// Export statistics for the specified training [sessionIds] to a PDF file
-  /// named `training_summary_filtered.pdf` located in the application documents
-  /// directory. [notes] should contain optional session notes keyed by session
-  /// id.
-  Future<String?> exportSessionsPdf(
-      List<int> sessionIds, Map<int, String> notes) async {
-    if (sessionIds.isEmpty) return null;
-
-    String durToStr(Duration d) => ExportUtils.durationString(d);
-
-    final regularFont = await pw.PdfGoogleFonts.robotoRegular();
-    final boldFont = await pw.PdfGoogleFonts.robotoBold();
-
-    final grouped = handsBySession();
-    final ids = List<int>.from(sessionIds)..sort();
-
-    final pdf = pw.Document();
-    pdf.addPage(
-      pw.MultiPage(
-        pageFormat: PdfPageFormat.a4,
-        build: (context) {
-          return [
-            for (final id in ids)
-              if (grouped[id] != null && grouped[id]!.isNotEmpty) ...[
-                () {
-                  final handsList = List<SavedHand>.from(grouped[id]!)
-                    ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-                  final stats = _sessionStats(handsList);
-                  final note = notes[id];
-
-                  return [
-                    pw.Text('Сессия $id',
-                        style: pw.TextStyle(font: boldFont, fontSize: 20)),
-                    pw.SizedBox(height: 4),
-                  pw.Text('Дата: ${formatDateTime(stats.end)}',
-                        style: pw.TextStyle(font: regularFont)),
-                  pw.Text('Длительность: ${durToStr(stats.duration)}',
-                        style: pw.TextStyle(font: regularFont)),
-                  pw.Text('Раздач: ${stats.count} • Верно: ${stats.correct} • Ошибки: ${stats.incorrect}',
-                        style: pw.TextStyle(font: regularFont)),
-                  if (stats.winrate != null)
-                    pw.Text('Winrate: ${stats.winrate!.toStringAsFixed(1)}%',
-                        style: pw.TextStyle(font: regularFont)),
-                    if (note != null && note.trim().isNotEmpty)
-                      pw.Text('Заметка: ${note.trim()}',
-                          style: pw.TextStyle(font: regularFont)),
-                    pw.SizedBox(height: 8),
-                    for (final hand in handsList)
-                      ...ExportUtils.handPdfWidgets(
-                        hand,
-                        regularFont,
-                        boldFont,
-                        titleSize: 16,
-                      ),
-                    pw.Divider(),
-                  ];
-                }(),
-              ],
-          ];
-        },
-      ),
-    );
-
-    final bytes = await pdf.save();
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/training_summary_filtered.pdf');
-    await file.writeAsBytes(bytes);
-    return file.path;
-  }
-
-  Future<String?> exportAllSessionsCsv(Map<int, String> notes) async {
-    if (hands.isEmpty) return null;
-
-    String durToStr(Duration d) => ExportUtils.durationString(d);
-
-    final grouped = handsBySession().entries.toList()
-      ..sort((a, b) => a.key.compareTo(b.key));
-    final rows = <List<dynamic>>[
-      ['Дата', 'Длительность', 'Раздач', 'Верно', 'EV', 'ICM']
-    ];
-
-    for (final entry in grouped) {
-      final list = List<SavedHand>.from(entry.value)
-        ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-      if (list.isEmpty) continue;
-      final stats = _sessionStats(list);
-      rows.add(ExportUtils.csvRow(
-        stats.end,
-        stats.duration,
-        stats.count,
-        stats.correct,
-        stats.evAvg,
-        stats.icmAvg,
-      ));
-    }
-
-    final csv = const ListToCsvConverter(fieldDelimiter: ';')
-        .convert(rows, eol: '\r\n');
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/training_summary.csv');
-    await file.writeAsString(csv);
-    return file.path;
-  }
-
-  Future<String?> exportSessionsCsv(
-      List<int> sessionIds, Map<int, String> notes) async {
-    if (sessionIds.isEmpty) return null;
-
-    String durToStr(Duration d) => ExportUtils.durationString(d);
-
-    final grouped = handsBySession();
-    final ids = List<int>.from(sessionIds)..sort();
-    final rows = <List<dynamic>>[
-      ['Дата', 'Длительность', 'Раздач', 'Верно', 'EV', 'ICM']
-    ];
-
-    for (final id in ids) {
-      final sessionHands = grouped[id];
-      if (sessionHands == null || sessionHands.isEmpty) continue;
-      final list = List<SavedHand>.from(sessionHands)
-        ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-      final stats = _sessionStats(list);
-      rows.add(ExportUtils.csvRow(
-        stats.end,
-        stats.duration,
-        stats.count,
-        stats.correct,
-        stats.evAvg,
-        stats.icmAvg,
-      ));
-    }
-
-    final csv = const ListToCsvConverter(fieldDelimiter: ';')
-        .convert(rows, eol: '\r\n');
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/training_summary_filtered.csv');
-    await file.writeAsString(csv);
-    return file.path;
-  }
-
-  /// Export hands belonging to [sessionId] to a Markdown file. The file
-  /// will be named `session_[id].md` and stored in the application documents
-  /// directory. Returns the created path or `null` if the session is empty.
-  Future<String?> exportSessionHandsMarkdown(int sessionId, {String? note}) async {
-    final sessionHands =
-        hands.where((h) => h.sessionId == sessionId).toList();
-    if (sessionHands.isEmpty) return null;
-    final buffer = StringBuffer();
-    if (note != null && note.trim().isNotEmpty) {
-      buffer.writeln(note.trim());
-      buffer.writeln();
-    }
-    for (final hand in sessionHands) {
-      buffer.write(ExportUtils.handMarkdown(hand));
-    }
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/session_$sessionId.md');
-    await file.writeAsString(buffer.toString());
-    return file.path;
-  }
-
-  /// Export hands belonging to [sessionId] to a PDF file. The file will be
-  /// named `session_[id].pdf` and stored in the application documents
-  /// directory. Returns the created path or `null` if the session is empty.
-  Future<String?> exportSessionHandsPdf(int sessionId, {String? note}) async {
-    final sessionHands =
-        hands.where((h) => h.sessionId == sessionId).toList();
-    if (sessionHands.isEmpty) return null;
-
-    final regularFont = await pw.PdfGoogleFonts.robotoRegular();
-    final boldFont = await pw.PdfGoogleFonts.robotoBold();
-
-    final pdf = pw.Document();
-    pdf.addPage(
-      pw.MultiPage(
-        pageFormat: PdfPageFormat.a4,
-        build: (context) {
-          return [
-            if (note != null && note.trim().isNotEmpty) ...[
-              pw.Text(note.trim(), style: pw.TextStyle(font: regularFont)),
-              pw.SizedBox(height: 12),
-            ],
-            for (final hand in sessionHands)
-              ...ExportUtils.handPdfWidgets(
-                hand,
-                regularFont,
-                boldFont,
-              )
-          ];
-        },
-      ),
-    );
-
-    final bytes = await pdf.save();
-    final dir = await getApplicationDocumentsDirectory();
-    final file = File('${dir.path}/session_$sessionId.pdf');
-    await file.writeAsBytes(bytes);
-    return file.path;
-  }
-
-  /// Export all sessions to a ZIP archive with Markdown or PDF files grouped
-  /// by session. Each session will be saved in a folder named by the session
-  /// date. Returns the archive path or `null` if there are no hands.
-  Future<String?> exportSessionsArchive({bool pdf = false}) async {
-    if (hands.isEmpty) return null;
-
-    String folder0(DateTime d) {
-      final y = d.year.toString().padLeft(4, '0');
-      final m = d.month.toString().padLeft(2, '0');
-      final day = d.day.toString().padLeft(2, '0');
-      return '$y-$m-$day';
-    }
-
-    final archive = Archive();
-    final grouped = handsBySession();
-    for (final entry in grouped.entries) {
-      final id = entry.key;
-      final list = List<SavedHand>.from(entry.value)
-        ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-      if (list.isEmpty) continue;
-      final folder = folder0(list.first.savedAt);
-      final path = pdf
-          ? await exportSessionHandsPdf(id)
-          : await exportSessionHandsMarkdown(id);
-      if (path == null) continue;
-      final file = File(path);
-      final data = await file.readAsBytes();
-      final ext = pdf ? 'pdf' : 'md';
-      final name = '$folder/session_$id.$ext';
-      archive.addFile(ArchiveFile(name, data.length, data));
-      await file.delete();
-    }
-
-    final bytes = ZipEncoder().encode(archive);
-    final dir = await getApplicationDocumentsDirectory();
-    final out =
-        File('${dir.path}/saved_hands_archive_${DateTime.now().millisecondsSinceEpoch}.zip');
-    await out.writeAsBytes(bytes);
-    return out.path;
-  }
 
   Future<SavedHand?> selectHand(BuildContext context) async {
     if (hands.isEmpty) return null;
@@ -956,127 +476,6 @@ class SavedHandManagerService extends ChangeNotifier {
     return selected;
   }
 
-  /// Group saved hands by their session identifier.
-  Map<int, List<SavedHand>> handsBySession() {
-    final Map<int, List<SavedHand>> grouped = {};
-    for (final hand in hands) {
-      grouped.putIfAbsent(hand.sessionId, () => []).add(hand);
-    }
-    return grouped;
-  }
-
-  List<SavedHand> filtered({String? tag, String? position, DateTimeRange? range}) {
-    return [
-      for (final h in hands)
-        if ((tag == null || h.tags.contains(tag)) &&
-            (position == null || h.heroPosition == position) &&
-            (range == null ||
-                (!h.date.isBefore(range.start) && !h.date.isAfter(range.end))))
-          h
-    ];
-  }
-
-  List<SavedHand> byTag(String tag) => filtered(tag: tag);
-
-  List<SavedHand> byPosition(String position) => filtered(position: position);
-
-  List<SavedHand> byDateRange(DateTimeRange range) => filtered(range: range);
-
-  /// Return the list of hands in the current error-free streak.
-  ///
-  /// Hands are returned in chronological order and only include
-  /// those with correct user action compared to GTO.
-  List<SavedHand> currentErrorFreeStreak() {
-    final result = <SavedHand>[];
-    for (int i = hands.length - 1; i >= 0; i--) {
-      final h = hands[i];
-      final expected = h.expectedAction?.trim().toLowerCase();
-      final gto = h.gtoAction?.trim().toLowerCase();
-      if (expected == null || gto == null || expected != gto) {
-        break;
-      }
-      result.insert(0, h);
-    }
-    return result;
-  }
-
-  /// Returns all completed error-free streaks of 5+ correct hands.
-  ///
-  /// Each streak is represented as a list of [SavedHand] objects in
-  /// chronological order. Only sequences terminated by a mistake are
-  /// included – the current ongoing streak (if any) is omitted.
-  List<List<SavedHand>> completedErrorFreeStreaks() {
-    final List<List<SavedHand>> streaks = [];
-    final List<SavedHand> current = [];
-
-    bool isCorrect(SavedHand h) {
-      final expected = h.expectedAction?.trim().toLowerCase();
-      final gto = h.gtoAction?.trim().toLowerCase();
-      return expected != null && gto != null && expected == gto;
-    }
-
-    for (final h in hands) {
-      if (isCorrect(h)) {
-        current.add(h);
-      } else {
-        if (current.length >= 5) {
-          streaks.add(List<SavedHand>.from(current));
-        }
-        current.clear();
-      }
-    }
-
-    // Only completed streaks are returned, so the trailing streak is ignored.
-    return streaks.reversed.toList();
-  }
-
-  List<SavedHand> filterByCategory(String category) {
-    final result = [
-      for (final h in hands)
-        if (h.category == category &&
-            h.expectedAction != null &&
-            h.gtoAction != null &&
-            h.expectedAction!.trim().toLowerCase() !=
-                h.gtoAction!.trim().toLowerCase())
-          h
-    ];
-    result.sort((a, b) => (b.evLoss ?? 0).compareTo(a.evLoss ?? 0));
-    return result;
-  }
-
-  bool hasSimilarMistakes(SavedHand hand) {
-    final cat = hand.category;
-    final pos = hand.heroPosition;
-    final stack = hand.stackSizes[hand.heroIndex];
-    if (cat == null || stack == null) return false;
-    return hands.any(
-      (h) =>
-          h != hand &&
-          h.category == cat &&
-          h.heroPosition == pos &&
-          h.stackSizes[h.heroIndex] == stack &&
-          h.expectedAction != null &&
-          h.gtoAction != null &&
-          h.expectedAction!.trim().toLowerCase() !=
-              h.gtoAction!.trim().toLowerCase(),
-    );
-  }
-
-  List<MapEntry<String, double>> getTopMistakeCategories({int limit = 3}) {
-    final map = <String, double>{};
-    for (final h in hands) {
-      final cat = h.category;
-      final exp = h.expectedAction;
-      final gto = h.gtoAction;
-      if (cat == null || cat.isEmpty) continue;
-      if (exp == null || gto == null) continue;
-      if (exp.trim().toLowerCase() == gto.trim().toLowerCase()) continue;
-      map[cat] = (map[cat] ?? 0) + (h.evLoss ?? 0);
-    }
-    final list = map.entries.toList()
-      ..sort((a, b) => b.value.compareTo(a.value));
-    return list.take(limit).toList();
-  }
 
   TrainingPackTemplate createPack(String name, List<SavedHand> selected) {
     HeroPosition pos(String s) => parseHeroPosition(s);
@@ -1113,50 +512,5 @@ class SavedHandManagerService extends ChangeNotifier {
       createdAt: DateTime.now(),
     );
     return tpl;
-  }
-
-  _SessionStats _sessionStats(List<SavedHand> hands) {
-    final list = List<SavedHand>.from(hands)
-      ..sort((a, b) => a.savedAt.compareTo(b.savedAt));
-    final start = list.first.savedAt;
-    final end = list.last.savedAt;
-    int correct = 0;
-    int incorrect = 0;
-    double evSum = 0;
-    double icmSum = 0;
-    int evCount = 0;
-    int icmCount = 0;
-    for (final h in list) {
-      final expected = h.expectedAction;
-      final gto = h.gtoAction;
-      if (expected != null && gto != null) {
-        if (expected.trim().toLowerCase() == gto.trim().toLowerCase()) {
-          correct++;
-        } else {
-          incorrect++;
-        }
-      }
-      final ev = h.heroEv;
-      if (ev != null) {
-        evSum += ev;
-        evCount++;
-      }
-      final icm = h.heroIcmEv;
-      if (icm != null) {
-        icmSum += icm;
-        icmCount++;
-      }
-    }
-    final evAvg = evCount > 0 ? evSum / evCount : null;
-    final icmAvg = icmCount > 0 ? icmSum / icmCount : null;
-    return _SessionStats(start, end, list.length, correct, incorrect,
-        evAvg: evAvg, icmAvg: icmAvg);
-  }
-
-  Future<void> _shareFile(File file) async {
-    if (Platform.isAndroid || Platform.isIOS) {
-      await Permission.storage.request();
-    }
-    await Share.shareXFiles([XFile(file.path)]);
   }
 }
